moved TaskTableColumns out of NonComp directory
